refactor: move DependantMap to another file
